backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
scikit-learn
joblib
motor
python-dotenv
pytest
requests

sse-starlette
